Friend sidebar css update
