+++ OurPlace/OurPlace/Models/User/SearchResult.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace OurPlace.Models.User
+{
+    public class SearchResult
+    {
+        public List<SearchUserModel> ModelList { get; set; }
+    }
+}
