+++ OurPlace/OurPlace/Views/Shared/_LoginPartial.cshtml
@@ -8,27 +8,27 @@
     @if (SignInManager.IsSignedIn(User))
     {
         <li class="nav-item">
-            <a class="nav-link text-dark" asp-controller="User" asp-action="Profile" asp-route-userId="@User.FindFirst(ClaimTypes.NameIdentifier).Value">My Profile</a>
+            <a class="nav-link text-light" asp-controller="User" asp-action="Profile" asp-route-userId="@User.FindFirst(ClaimTypes.NameIdentifier).Value">My Profile</a>
         </li>
         <li class="nav-item">
-            <a class="nav-link text-dark" id="friendBtn" onclick="Toggle()">Friends</a>
+            <a class="nav-link text-light" id="friendBtn" onclick="Toggle()">Friends</a>
         </li>
         <li class="nav-item">
-            <a class="nav-link text-dark" asp-controller="Notification" asp-action="Overview" asp-route-userId="@User.FindFirst(ClaimTypes.NameIdentifier).Value">Notifications</a>
+            <a class="nav-link text-light" asp-controller="Notification" asp-action="Overview" asp-route-userId="@User.FindFirst(ClaimTypes.NameIdentifier).Value">Notifications</a>
         </li>
         <li class="nav-item">
             <form class="form-inline" asp-area="Identity" asp-page="/Account/Logout" asp-route-returnUrl="@Url.Action("HomePage", "Home", new { area = "" })">
-                <button type="submit" class="nav-link btn btn-link text-dark">Logout</button>
+                <button type="submit" class="nav-link btn btn-link text-light">Logout</button>
             </form>
         </li>
     }
     else
     {
         <li class="nav-item">
-            <a class="nav-link text-dark" asp-area="Identity" asp-page="/Account/Register">Register</a>
+            <a class="nav-link text-light" asp-area="Identity" asp-page="/Account/Register">Register</a>
         </li>
         <li class="nav-item">
-            <a class="nav-link text-dark" asp-area="Identity" asp-page="/Account/Login">Login</a>
+            <a class="nav-link text-light" asp-area="Identity" asp-page="/Account/Login">Login</a>
         </li>
     }
 </ul>
