+++ OurPlace/OurPlace/wwwroot/css/Friend.css
@@ -9,7 +9,10 @@
 }
 
 .friendSidebard {
-    width: 270px;
+    margin-top: -1rem;
+    width: 200px;
+    height: 90%;
+    background-color: #343a40;
     transition: left 0.6s;
     position: absolute;
     z-index: 20;
@@ -17,7 +20,7 @@
 
 .sidebarHeader {
     padding: 20px;
-    background-color: rgb(79, 79, 79);
+    background-color: #218838;
     text-align: center;
     margin-bottom: 0 !important;
     color: whitesmoke;
@@ -25,11 +28,11 @@
 
 .friend {
     color: whitesmoke;
-    padding: 20px;
-    background-color: rgb(35, 35, 35);
+    padding: 15px;
+    background-color: #343a40;
     text-align: center;
     font-size: 1.2em;
-    border-top: 1px solid rgb(154, 154, 154);
+    border-bottom: 1px solid rgb(154, 154, 154);
     cursor: pointer;
 }
     .friend:hover {
